code refactoring (move types to its own folders, cut start page for small components)
