+++ src/components/CustomDivider/types/index.ts
@@ -1,0 +1,4 @@
+export type DividerPropsType = {
+  position: string;
+  height?: number;
+};
