+++ src/components/Header/types/index.ts
@@ -1,0 +1,7 @@
+export type HeaderPropsType = {
+  showBack?: boolean;
+  showUpdate?: boolean;
+  showMenu?: boolean;
+  isLoading?: boolean;
+  refreshNews?: () => void;
+};
