+++ src/components/Logo/types/index.ts
@@ -1,0 +1,5 @@
+export type LogoPropsType = {
+  size: number;
+  testId?: string;
+  color: string;
+};
