+++ src/components/NestedCommentsWrapper/components/NestedCommentItem/index.tsx
@@ -2,22 +2,16 @@ import { FC } from "react";
 import { Box, Typography } from "@mui/material";
 
 import { CustomDivider } from "components/CustomDivider";
-import { CommentType } from "types/commonTypes";
 import { DateUtil } from "utiles/DateUtil/DateUtil";
 import { DIVIDER_TYPE } from "constants/dividerTypes";
 
+import { WrapperPropsType } from "components/NestedCommentsWrapper/types";
 import { DeletedNestedComment } from "../DeletedNestedComment";
 import { styles } from "./styles";
 import { NestedCommentsWrapper } from "components/NestedCommentsWrapper";
 import { TEST_ID } from "constants/testIds";
 
-type NestedCommentItemPropsType = {
-  comment: CommentType;
-};
-
-export const NestedCommentItem: FC<NestedCommentItemPropsType> = ({
-  comment,
-}) => {
+export const NestedCommentItem: FC<WrapperPropsType> = ({ comment }) => {
   if (comment.deleted) return <DeletedNestedComment />;
 
   return (
