+++ src/components/NestedCommentsWrapper/types/index.ts
@@ -1,0 +1,9 @@
+import { CommentType } from "types/commonTypes";
+
+export type WrapperPropsType = {
+  comment: CommentType;
+};
+
+export type NestedCommentsPropsType = {
+  commentId: number;
+};
