+++ src/components/News/types/index.ts
@@ -1,0 +1,5 @@
+import { StoryType } from "types/commonTypes";
+
+export type NewsItemPropsType = {
+  item: StoryType;
+};
