+++ src/components/ProgressLoader/types/index.ts
@@ -1,0 +1,5 @@
+export type LoaderTypeProps = {
+  fullScreen?: boolean;
+  centerInParent?: boolean;
+  size?: number;
+};
