+++ src/components/StartPage/components/QuoteSkeleton/styles.ts
@@ -1,0 +1,9 @@
+export const styles = {
+  wrapper: {
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "flex-end",
+    alignItems: "flex-end",
+    marginTop: "4rem",
+  },
+};
