+++ src/components/StartPage/constants/getStoriesTypeText.ts
@@ -1,6 +1,6 @@
 import { STORIES_SHOWED_TYPE } from "constants/storiesTypes";
 
-export const storiesTypeConfig = {
+export const getStoriesTypeText = {
   [STORIES_SHOWED_TYPE.ASK]: "Ask Stories",
   [STORIES_SHOWED_TYPE.BEST]: "Best Stories",
   [STORIES_SHOWED_TYPE.JOB]: "Job Stories",
