+++ src/components/StartPage/index.tsx
@@ -1,53 +1,34 @@
-import { useState, MouseEvent, useEffect } from "react";
+import { useState, useEffect } from "react";
 import { useAppDispatch } from "store/hooks";
-import { Box, Button, Typography } from "@mui/material";
-import ArrowDropDownIcon from "@mui/icons-material/ArrowDropDown";
-import ArrowDropUpIcon from "@mui/icons-material/ArrowDropUp";
-import { useNavigate } from "react-router-dom";
-import { useSelector } from "react-redux";
+import { Box } from "@mui/material";
+
 import { Logo } from "components/Logo";
-import { StoriesTypeSwitcher } from "components/StoriesTypeSwitcher";
 import { CustomDivider } from "components/CustomDivider";
-import { DIVIDER_TYPE } from "constants/dividerTypes";
-import { SCREEN_ROUTES } from "routes/constants";
-import { getShowedStoryType } from "store/hackerNews/selectors/getShowedStoryType";
 import { getRandomQuote } from "store/hackerNews/actions";
+import { DIVIDER_TYPE } from "constants/dividerTypes";
 import { TEST_ID } from "constants/testIds";
 import { orangePrimary } from "constants/colors";
-import { storiesTypeConfig } from "./constants/storiesTypeConfig";
-import { styles } from "./styles";
 
-type QuoteType = {
-  text: string;
-  author: string;
-};
+import { Footer } from "./components/Footer";
+import { QuoteType } from "./types";
+import { styles } from "./styles";
+import { ActionsBlock } from "./components/ActionsBlock";
 
 function StartPage() {
-  const navigate = useNavigate();
   const dispatch = useAppDispatch();
-  const [switcherOpen, setSwitcherOpen] = useState<HTMLElement | null>(null);
   const [quote, setQuote] = useState<QuoteType>({ text: "", author: "" });
-  const storiesShowedType = useSelector(getShowedStoryType);
-
-  const isSwitcherOpen = Boolean(switcherOpen);
-
-  const goToNews = () => {
-    navigate(SCREEN_ROUTES.NEWS_LIST);
-  };
-
-  const closeSwitcher = () => {
-    setSwitcherOpen(null);
-  };
-
-  const openSwitcher = (e: MouseEvent<HTMLElement>) => {
-    setSwitcherOpen(e.currentTarget);
-  };
+  const [isLoading, setIsLoading] = useState<boolean>(false);
 
   useEffect(() => {
+    setQuote({ text: "", author: "" });
+    setIsLoading(true);
     dispatch(getRandomQuote())
       .unwrap()
       .then((res) => {
-        setQuote(res);
+        setQuote({ text: `"${res.text}"`, author: `(c) ${res.author}` });
+      })
+      .finally(() => {
+        setTimeout(() => setIsLoading(false), 300);
       });
   }, []);
 
@@ -65,39 +46,9 @@ function StartPage() {
           />
         </Box>
         <CustomDivider position={DIVIDER_TYPE.VERTICAL} height={320} />
-        <Box sx={styles.contentWrapper}>
-          <Box sx={styles.selectorWrapper}>
-            <Typography sx={styles.selectorHelper}>
-              Which news do you want to read?
-            </Typography>
-            <Box sx={styles.switcherButton} onClick={openSwitcher}>
-              <Typography sx={styles.storiesType}>
-                {storiesShowedType}
-              </Typography>
-              {isSwitcherOpen ? <ArrowDropUpIcon /> : <ArrowDropDownIcon />}
-            </Box>
-          </Box>
-          <Button
-            onClick={goToNews}
-            sx={styles.showNewsButton}
-            data-testid={TEST_ID.START_PAGE.GO_TO_NEWS_BUTTON}
-          >
-            Go to {storiesTypeConfig[storiesShowedType]}
-          </Button>
-          {isSwitcherOpen && (
-            <StoriesTypeSwitcher
-              open={isSwitcherOpen}
-              switcherOpen={switcherOpen}
-              closeSwitcher={closeSwitcher}
-              fullWidth
-            />
-          )}
-        </Box>
+        <ActionsBlock setQuote={setQuote} />
       </Box>
-      <Box sx={styles.bottomContent}>
-        <Typography sx={styles.quoteTitle}>"{quote.text}"</Typography>
-        <Typography sx={styles.quoteAuthor}>{`(c) ${quote.author}`}</Typography>
-      </Box>
+      <Footer isLoading={isLoading} quote={quote} />
     </Box>
   );
 }
