+++ src/components/StoriesTypeSwitcher/types/index.ts
@@ -1,0 +1,6 @@
+export type MenuPropsType = {
+  open: boolean;
+  switcherOpen: HTMLElement | null;
+  closeSwitcher: () => void;
+  fullWidth?: boolean;
+};
