+++ src/components/StoryPage/types/index.ts
@@ -1,0 +1,14 @@
+import { CommentType } from "types/commonTypes";
+
+export type CommentItemPropsType = {
+  comment: CommentType;
+};
+
+export type CommentsPropsType = {
+  storyId: number | undefined;
+  isLoading: boolean;
+};
+
+export type StoryPageHeaderPropsType = {
+  isLoading: boolean;
+};
